+++ profile.html
@@ -9,6 +9,11 @@
 </head>
 <body>
     <div id="navbar"></div>
+    <div id="location-menu"></div>
+    <div id="cart"></div>
+    <!-- <div>
+        
+    </div> -->
     <!-- import nav bar here -->
 
 <div class="profile-parent">
@@ -118,3 +123,4 @@
 </div>
 </body>
 </html>
+<script type="module" src="./scripts/profile.js"></script>
