+++ scripts/fitness.js
@@ -1,3 +1,6 @@
+
+
+
 const productContainers = document.querySelectorAll(".product-container")
 const nxtBtn = document.querySelectorAll(".nxt-btn")
 const preBtn = document.querySelectorAll(".pre-btn")
@@ -14,3 +17,14 @@ productContainers.forEach((item,i) => {
         item.scrollLeft += containerwidth;
     })
 })
+
+const sliderMain = document.getElementById('slider-main');
+const item = sliderMain.getElementsByClassName('item');
+
+function next() {
+    sliderMain.append(item[0]);
+}
+function prev() {
+    sliderMain.prepend(item[item.length-1]);
+}
+
