+++ scripts/profile.js
@@ -1,0 +1,9 @@
+import {navbar, locateDiv, navbarFunc, cartDiv, showCartItems} from "../component/navbar.js";
+
+    let navbarDiv = document.getElementById("navbar");
+    navbarDiv.innerHTML = navbar();
+    navbarDiv.style.backgroundColor = 'black';
+    document.getElementById("location-menu").innerHTML = locateDiv();
+    document.getElementById("cart").innerHTML = cartDiv();
+    navbarFunc();
+    showCartItems();
