--- slider.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./styles/slider.css">
-    <link rel="stylesheet" href="./fitness.html">
-    <title>slider page</title>
-</head>
-<body>
-<div id="main-container">
-<div><h1 id="h1tag">Conters HSR Layout</h1></div>
-<div class="product">
-    <div class="product-container">
-        <button class="pre-btn"><img src="https://www.iconpacks.net/icons/2/free-arrow-left-icon-3099-thumb.png" alt=""></button>
-        <!-- 1 -->
-        <div class="product-card">
-            <div class="product-image">
-                <img class="product-thumb" src="https://cdn-images.cure.fit/www-curefit-com/image/upload/w_300,fl_progressive,f_auto,q_auto:eco/dpr_2/image/icons/cult/gyms_default.jpg" alt="">
-                <button class="card-btn">TRY FOR FREE</button>
-            </div>
-            <div class="product-info">
-                <h1 class="product-brand">FUERZA</h1>
-                <p class="product-short-description">HSR Layout.Gym</p>
-                <p class="product-short-description">Unlimited access starting at ₹714/mo</p>
-            </div>
-        </div>
-        <!-- 2 -->
-        <div class="product-card">
-            <div class="product-image">
-                <img class="product-thumb" src="https://cdn-images.cure.fit/www-curefit-com/image/upload/w_300,fl_progressive,f_auto,q_auto:eco/dpr_2/image/gymfit/centers/CENTER45/e4c3bbf7-4e7e-4e0d-90c6-a58fb6b4b1c0.jpg" alt="">
-                <button class="card-btn">TRY FOR FREE</button>
-            </div>
-            <div class="product-info">
-                <h1 class="product-brand">ZEST FITNESS</h1>
-                <p class="product-short-description">HSR Layout.Gym</p>
-                <p class="product-short-description">Unlimited access with cultpass ELITE & PRO</p>
-            </div>
-        </div>
-        <!-- 3 -->
-        <div class="product-card">
-            <div class="product-image">
-                <img class="product-thumb" src="https://cdn-images.cure.fit/www-curefit-com/image/upload/w_300,fl_progressive,f_auto,q_auto:eco/dpr_2/image/gymfit/centers/prod/CENTER134/7f22613a-440a-4281-a4fa-4a62ed9a7f96_DSC1213.jpg" alt="">
-                <button class="card-btn">TRY FOR FREE</button>
-            </div>
-            <div class="product-info">
-                <h1 class="product-brand">PLAY ON FITNESS</h1>
-                <p class="product-short-description">HSR Layout.Gym</p>
-                <p class="product-short-description">Unlimited access with cultpass ELITE & PRO</p>
-            </div>
-        </div>
-        <!-- 4 -->
-        <div class="product-card">
-            <div class="product-image">
-                <img class="product-thumb" src="https://cdn-images.cure.fit/www-curefit-com/image/upload/w_300,fl_progressive,f_auto,q_auto:eco/dpr_2/v1/cult-media/v2web/centers/3_Cult_HSR_PRODUCT_BNR_2018-08-10T12:18:41.704Z.png" alt="">
-                <button class="card-btn">TRY FOR FREE</button>
-            </div>
-            <div class="product-info">
-                <h1 class="product-brand">CULT HSR</h1>
-                <p class="product-short-description">HSR Layout.Gym</p>
-                <p class="product-short-description">Unlimited access starting at ₹1100/mo</p>
-            </div>
-        </div>
-        <!-- 5 -->
-        <div class="product-card">
-            <div class="product-image">
-                <img class="product-thumb" src="https://cdn-images.cure.fit/www-curefit-com/image/upload/w_300,fl_progressive,f_auto,q_auto:eco/dpr_2/v1/cult-media/v2web/centers/220_Cult_HSR%2024th%20Main%20_PRODUCT_BNR_2021-11-09T09:52:29.369Z.png" alt="">
-                <button class="card-btn">TRY FOR FREE</button>
-            </div>
-            <div class="product-info">
-                <h1 class="product-brand">CULT HSR 24TH MAIN</h1>
-                <p class="product-short-description">HSR Layout.Gym</p>
-                <p class="product-short-description">Unlimited access starting at ₹1100/mo</p>
-            </div>
-        </div>
-        <!-- 6 -->
-        <div class="product-card">
-            <div class="product-image">
-                <img class="product-thumb" src="https://cdn-images.cure.fit/www-curefit-com/image/upload/w_300,fl_progressive,f_auto,q_auto:eco/dpr_2/image/icons/cult/no_gyms.png" alt="">
-                <button class="card-btn">SEE MORE</button>
-            </div>
-            <div class="product-info">
-                <h1 class="product-brand">MORE CENTERS</h1>
-                <p class="product-short-description">Click on see more to expand your search</p>
-            </div>
-        </div>
-        <button class="nxt-btn"><img src="https://static.thenounproject.com/png/1247048-200.png" alt=""></button>    
-    </div>
-</div>
-</div>
-</body>
-</html>
-<script src="./scripts/slider.js"></script>
