--- styles/slider.css
@@ -1,119 +1,0 @@
-#main-container{
-    height: 550px;
-    background-color: #14161b;
-}
-#h1tag{
-    color: #fff;
-    margin-left: 75px;
-    padding-top: 25px;
-}
-.pre-btn img{
-    width: 15px;
-    height: 30px;
-    border: none;
-}
-.nxt-btn img{
-    width: 15px;
-    height: 30px;
-    border: none;
-}
-.product{
-    background-color: #14161b;
-    position: relative;
-    overflow: hidden;
-    padding: 5px;
-    scroll-behavior: inherit;
-}
-.product-container{
-    padding: 0 3vw;
-    display: flex;
-    overflow-x: auto;
-}
-.product-card {
-    height: 450px;
-    width: 550px;
-    margin-right:100px;
-    flex: 0 0 auto;
-    background-color: #44474b;
-}
-.product-image{
-    position: relative;
-    width: 100%;
-    height: 350px;
-}
-.product-thumb{
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-.card-btn{
-    position: absolute;
-    bottom: 10px;
-    left: 50%;
-    transform: translateX(-50%);
-    padding: 10px;
-    width: 30%;
-    text-transform: capitalize;
-    border: none;
-    outline: none;
-    border-radius: 5px;
-    transition: 0.5s;
-    cursor: pointer;
-    opacity: 0;
-    background-color: #44474b;
-}
-.product-card:hover .card-btn{
-    opacity: 1;
-}
-.card-btn:hover{
-    background-color: #44474b;
-    color: #fff;
-    font-size: 15px;
-}
-.product-info{
-    width: 100%;
-    height: 100%;
-    padding-top: 0px;
-    padding-left: 15px;
-}
-.product-brand{
-    text-transform: uppercase;
-    padding-top: 0px;
-    margin-top: 0px;
-    color: #fff;
-}
-.product-short-description{
-    width: 100%;
-    height: 20px;
-    line-height: 20px;
-    overflow: hidden;
-    opacity: 0.5;
-    text-transform: capitalize;
-    margin: 0px 0;
-    color: black;
-    color: #fff;
-}
-.pre-btn
-.nxt-btn{
-    border: none;
-    width: 10vw;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    display: flex;
-    justify-content: center;
-    cursor: pointer;
-    z-index: 9;
-}
-.pre-btn{
-    left: 0;
-    width: 40px;
-    height: 50px;
-    margin-top: 200px;
-}
-.nxt-btn{
-    right: 0;
-    width: 40px;
-    height: 50px;
-    margin-top: 200px;
-}
