--- styles/trainer.css
@@ -1,76 +1,0 @@
-#trainer{
-    height: 150px;
-    width: 100%;
-    text-align: center;
-    padding-top: 65px;
-    background-color: #15171c;       
-}
-#center > h3{
-    color: rgb(141, 143, 145);
-    font-size: 25px;
-    font-weight: 1000;
-    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
-}
-#center > h1{
-    color: aliceblue;
-    font-size: 40px;
-    font-weight: 1000;
-    font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
-}
-*{
-    padding: 0;
-    margin: 0;
-}
-.explore-slider{
-    background-color: #1c2734;
-    width: 100%;
-    height: 380px;
-    overflow: hidden;
-    position: relative;
-}
-span{
-    position: absolute;
-    top: 40%;
-    font-size: 60px;
-    font-weight: bold;
-    color: #ffff;
-    cursor: pointer;
-    left: .5%;
-    z-index: 1;
-    margin-left: 7%;
-    background-color: #1c2734;
-}
-span:nth-of-type(2){
-    left: 84%;
-    text-align: right;
-}
-.slider-main{
-    background-color: #1c2734;
-    width: 80%;
-    height: 350px;
-    background: #f4f4f4;
-    margin: 10px auto;
-    display: flex;
-    justify-content: left;
-    overflow-x: auto;
-}
-.slider-main::-webkit-scrollbar{
-    visibility: hidden;
-}
-.slider-main > .item{
-    border: 1px solid green;
-    min-width: 250px;
-    height: 350px;
-    left: 0;
-    text-align: center;
-    font-size: 100px;
-    font-weight: bold;
-    margin: 10px 13px;
-    color: white;
-    background: purple;
-    transition: 0.5s;
-}
-.item > img{
-    width: 100%;
-    height: 100%;
-}
